Cap stat_zero time, don't save it as a negative number.
Recovery is now limited to at most 200 turns. Previously it was
unlimited, but after a save and restore it would either take 127 or
fewer turns, or take forever as the time was restored as a negative
number.  Besides the cap on recovery time, we marshall as an
unsigned byte, not signed.

The cap could be increased, but anything larger than 255 will
require a minor tag and a change to the marshalling format.

+++ crawl-ref/source/player.h
@@ -63,7 +63,7 @@ public:
 
   FixedVector<int8_t, NUM_STATS> stat_loss;
   FixedVector<int8_t, NUM_STATS> base_stats;
-  FixedVector<int, NUM_STATS> stat_zero;
+  FixedVector<uint8_t, NUM_STATS> stat_zero;
 
   int hunger;
   int disease;
